+++ src/main/java/com/example/demo/service/UserPermissionServiceImpl.java
@@ -4,6 +4,8 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,6 +17,8 @@ public class UserPermissionServiceImpl implements UserPermissionService{
 
     @Autowired
     private UserPermissionRepository userPermissionRepository;
+
+    private static final Logger logger =LoggerFactory.getLogger(UserPermissionServiceImpl.class);
 
     @Override
     public List<UserPermission> getAllUserPermissions() {
@@ -37,14 +41,24 @@ public class UserPermissionServiceImpl implements UserPermissionService{
 
     @Override
     public UserPermission updateUserPermission(long id, UserPermission userPermission) {
-        UserPermission exixstingUserPermission = getUserPermissionById(id);
-        exixstingUserPermission.setUserPermissionName(userPermission.getUserPermissionName());
-        return userPermissionRepository.save(exixstingUserPermission);
+        UserPermission existingUserPermission = getUserPermissionById(id);
+
+        if((existingUserPermission.getUserPermissionName().equals(userPermission.getUserPermissionName()))){
+            throw new IllegalArgumentException("Permission with name" + userPermission.getUserPermissionName() + " already exists");
+        }
+
+        existingUserPermission.setUserPermissionName(userPermission.getUserPermissionName());
+        return userPermissionRepository.save(existingUserPermission);
     }
 
     @Override
     public void deleteRole(long id) {
+        Optional<UserPermission> existingUserPermission = userPermissionRepository.findById(id);
+        
+        if(!existingUserPermission.isPresent()){
+            throw new IllegalArgumentException("Permission is not found with id: " + id);
+        }
         userPermissionRepository.deleteById(id);
+        logger.info("Permission with id {} was deleted.", id);
     }
-    
 }
